checking if a firm is registered and then redirecing to login if true else directs to registration

+++ src/auth/functions/registered.js
@@ -8,10 +8,13 @@ registered(){
     let id = localStorage.getItem('firm_id');
     if(!_.isEmpty(id)){
         return this.isRegistered = true
+        
     }
-    else {
+    else if(_.isEmpty(id)){
         localStorage.clear()
         return this.isRegistered = false
+        
     }
  }
 }
+console.log(auth.registered())
